requirements.txt
pytest==2.9.0
pytest-allure-adaptor==1.7.6
pytest-xdist==1.15.0
selenium==3.0.2
